flaskr/requirements.txt
Flask==2.2.3
Flask-MySQL==1.5.2
PyMySQL==1.0.3
jsonschema==4.17.3
pycryptodome==3.17